Layers form: split reducer and initial state
