+++ components/admin/layers/form/interactions/index.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import * as actions from './interactions-actions';
 import * as reducers from './interactions-reducer';
-import { initialState } from './interactions-reducer';
+import initialState from './interactions-default-state';
 
 import InteractionsComponent from './interactions-component';
