+++ components/admin/layers/form/interactions/interactions-default-state.js
@@ -1,0 +1,5 @@
+export default {
+  added: [],
+  available: [],
+  loading: false
+};
