+++ components/admin/layers/form/layer-preview/layer-preview-default-state.js
@@ -1,0 +1,6 @@
+export default {
+  interaction: {},
+  interactionLatLng: null,
+  interactionSelected: null,
+  layerGroups: []
+};
